main: add frame timing
Fixes issue #10 whereby gameplay would run at the speed of the monitor
refresh rate. A delay is automatically calculated on each frame to slow
down the gameplay if required. Note that this does not speed up the
gameplay if the monitor has a refresh rate lower than 59 Hz. In such
cases, VSYNC should just be disabled in the source code.

Signed-off-by: Mahyar Koshkouei <mk@deltabeard.com>